Update to 2.2.6
Also added an explicit dependency on libayatana-appindicator, as it's
necessary for snx-rs-gui to run. While I was at it, I moved the
dependencies on systemd and iproute2 to runtime dependencies, as they're
not needed to build the package. However, iproute2 is neeeded for tests,
so I've added it to checkdepends.